@@ -1,14 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:kakao_t_ui_exam/ui/counter/counter_screen.dart';
 import 'package:kakao_t_ui_exam/ui/home/home_screen.dart';
-import 'package:kakao_t_ui_exam/ui/kakao_t/kakao_t_screen.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   // This widget is the root of your application.
   @override
