@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:kakao_t_ui_exam/ui/bmi_calc/bmi_calc_result_screen.dart';
 
 class BmiCalcHomeScreen extends StatefulWidget {
-  const BmiCalcHomeScreen({Key key}) : super(key: key);
+  const BmiCalcHomeScreen({Key? key}) : super(key: key);
 
   @override
   State<BmiCalcHomeScreen> createState() => _BmiCalcHomeScreenState();
@@ -15,9 +15,9 @@ class _BmiCalcHomeScreenState extends State<BmiCalcHomeScreen> {
 
   @override
   void dispose() {
-    super.dispose();
     _heightController.dispose();
     _weightController.dispose();
+    super.dispose();
   }
 
   @override
@@ -37,10 +37,10 @@ class _BmiCalcHomeScreenState extends State<BmiCalcHomeScreen> {
                   border: OutlineInputBorder(),
                   hintText: '키',
                 ),
-              //  keyboardType: TextInputType.number,
+               keyboardType: TextInputType.number,
                 controller: _heightController,
                 validator: (value) {
-                  if (value.trim().isEmpty) {
+                  if (value!.trim().isEmpty) {
                     return '키를 입력하세요';
                   }
                   return null;
@@ -54,10 +54,10 @@ class _BmiCalcHomeScreenState extends State<BmiCalcHomeScreen> {
                   border: OutlineInputBorder(),
                   hintText: '몸무게',
                 ),
-               // keyboardType: TextInputType.number,
+               keyboardType: TextInputType.number,
                 controller: _weightController,
                 validator: (value) {
-                  if (value.trim().isEmpty) {
+                  if (value != null && value.trim().isEmpty) {
                     return '몸무게를 입력하세요';
                   }
                   return null;
@@ -68,7 +68,11 @@ class _BmiCalcHomeScreenState extends State<BmiCalcHomeScreen> {
                 alignment: Alignment.centerRight,
                 child: ElevatedButton(
                   onPressed: () {
-                    if (_formKey.currentState.validate()) {
+                    // if(_formKey.currentState == null) {
+                    //   return;
+                    // }
+                    // if(_formKey.currentState!.validate()) {
+                    if (_formKey.currentState?.validate() ?? false) {
                       Navigator.push(
                         context,
                         MaterialPageRoute(
