@@ -8,7 +8,7 @@ import 'package:kakao_t_ui_exam/ui/kakao_t/detail_screen.dart';
 import 'package:kakao_t_ui_exam/ui/kakao_t/widgets/menu_widget.dart';
 
 class KakaoTScreen extends StatefulWidget {
-  const KakaoTScreen({Key key}) : super(key: key);
+  const KakaoTScreen({Key? key}) : super(key: key);
 
   @override
   State<KakaoTScreen> createState() => _KakaoTScreenState();
